@@ -17,61 +17,64 @@
     },
     "require": {
         "php": ">=7.1",
-        "symfony/symfony": "2.8.*",
+        "symfony/symfony": "3.4.*",
+
+        "symfony/swiftmailer-bundle": "^3.2",
+        "symfony/webpack-encore-pack": "^1.0",
+        "symfony/monolog-bundle": "^3.3",
+
+        "friendsofsymfony/user-bundle": "^2.1",
+        "friendsofsymfony/jsrouting-bundle": "^2.2",
+        "sonata-project/user-bundle": "^4.2",
+        "sonata-project/admin-bundle": "^3.37",
+        "sonata-project/doctrine-orm-admin-bundle": "^3.6",
+        "sonata-project/datagrid-bundle": "^2.3",
+        "fr3d/ldap-bundle": "^3.0",
+
+        "sensio/distribution-bundle": "^5.0",
+        "sensio/framework-extra-bundle": "^5.2",
+        "incenteev/composer-parameter-handler": "^2.1",
 
         "ext-json": "^1.5",
         "facebook/graph-sdk": "^5.6",
         "symfony/assetic-bundle": "^2.8",
         "google/apiclient": "^2.2",
-        "symfony/swiftmailer-bundle": "^3.2",
-        "symfony/webpack-encore-pack": "^1.0",
-        "symfony/monolog-bundle": "^3.3",
+        "willdurand/geocoder-bundle": "^5.1",
+        "eightpoints/guzzle-bundle": "^7.3",
 
-        "doctrine/orm": "~2.2,>=2.2.3",
-        "doctrine/doctrine-bundle": "~1.2",
-        "doctrine/doctrine-fixtures-bundle": "^2.4",
-        "doctrine/data-fixtures": "dev-master",
-        "doctrine/migrations": "*@dev",
-        "doctrine/doctrine-migrations-bundle": "1.0.*",
-
-        "sensio/distribution-bundle": "~5.0",
-        "sensio/framework-extra-bundle": "~3.0",
-        "incenteev/composer-parameter-handler": "~2.0",
+        "doctrine/doctrine-bundle": "^1.9",
+        "doctrine/orm": "^2.6",
+        "doctrine/doctrine-fixtures-bundle": "^3.0",
+        "doctrine/data-fixtures": "^1.3",
+        "doctrine/migrations": "^1.8",
+        "doctrine/doctrine-migrations-bundle": "^1.3",
 
         "hwi/oauth-bundle": "*@dev",
         "knplabs/knp-paginator-bundle": "dev-master",
         "braincrafted/bootstrap-bundle": "dev-master",
         "twitter/bootstrap": "2.3.*",
-        "friendsofsymfony/jsrouting-bundle": "~1.5",
-        "fr3d/ldap-bundle": "2.0.*@dev",
+
         "ircmaxell/password-compat": "~1.0",
         "liip/theme-bundle": "dev-master",
-        "willdurand/geocoder-bundle": "@stable",
-        "eightpoints/guzzle-bundle": "^5.1",
         "twbs/bootstrap": "^3.3",
-        "behat/mink": "1.7",
+
         "dmore/behat-chrome-extension": "^1.2",
         "symfony/polyfill-iconv": "^1.8",
-        
-        "friendsofsymfony/user-bundle": "^2.1",
-        "sonata-project/user-bundle": "^4.2",
-        "sonata-project/admin-bundle": "^3.37",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.6",
-        "sonata-project/datagrid-bundle": "^2.3"
+        "ext-SimpleXML": "^7.1",
+        "friendsofsymfony/rest-bundle": "^2.3"
     },
     "require-dev": {
-        "sensio/generator-bundle": "2.3.*",
-        "phpspec/phpspec": "~2.5@dev",
-        "henrikbjorn/phpspec-code-coverage": "1.0.*@dev",
-        "phpunit/phpunit": "3.7.*",
-        "bossa/phpspec2-expect": "dev-master",
-        "behat/behat": "~3.0.6@dev",
-        "behat/symfony2-extension": "2.0.*@dev",
-        "behat/mink-extension": "*@dev",
-        "behat/mink-browserkit-driver": "*@dev",
-        "behat/mink-goutte-driver": "*@dev",
-        "behat/mink-selenium2-driver": "*@dev",
-        "sensiolabs/behat-page-object-extension": "*@dev",
+        "behat/behat": "^3.5",
+        "behat/symfony2-extension": "^2.1",
+        "behat/mink-extension": "^2.3",
+        "behat/mink-browserkit-driver": "^1.3",
+        "behat/mink-goutte-driver": "^1.2",
+        "behat/mink-selenium2-driver": "^1.3",
+        "symfony/phpunit-bridge": "*",
+        "phpunit/phpunit": "^7.3",
+        "phpspec/phpspec": "^5.0",
+        "sensio/generator-bundle": "^3.1",
+        "sensiolabs/behat-page-object-extension": "^2.1",
         "vipsoft/doctrine-data-fixtures-extension": "3.0.*@dev"
     },
     "scripts": {
