@@ -4,7 +4,7 @@ namespace Tests\CatrobatCodeParserTests;
 
 use Catrobat\AppBundle\Services\CatrobatCodeParser\ParsedScene;
 
-class ParsedObjectsContainerTest extends \PHPUnit_Framework_TestCase
+class ParsedObjectsContainerTest extends \PHPUnit\Framework\TestCase
 {
     protected $container;
 
@@ -66,10 +66,11 @@ class ParsedObjectsContainerTest extends \PHPUnit_Framework_TestCase
      */
     public function mustThrowExceptionIfCorruptedGroup()
     {
-        $this->setExpectedException('\Exception');
+        $this->expectExceptionMessage('\Exception');
 
         $xml_properties = simplexml_load_file(__DIR__
             . '/Resources/FaultyPrograms/CorruptedGroupFaultyProgram/code.xml');
+        
         new ParsedScene($xml_properties->xpath('//scene')[0]);
     }
 }
