@@ -7,7 +7,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 class LogoutController extends Controller
 {
     public function logoutAction() {
-        $this->get('security.context')->setToken(null);
+        $this->get('security.token_storage')->setToken(null);
 
         return $this->redirect($this->generateUrl('index'));
     }
