@@ -4,11 +4,11 @@ namespace PhpSpec\CatrobatExtension;
 
 use PhpSpec\ServiceContainer;
 
-class Extension implements \PhpSpec\Extension\ExtensionInterface
+class Extension implements \PhpSpec\Extension
 {
-    public function load(ServiceContainer $container)
+    public function load(ServiceContainer $container, array $params = [])
     {
-        $container->setShared('event_dispatcher.listeners.catrobat', function ($container) {
+      $container->define('event_dispatcher.listeners.catrobat', function ($container) {
       $listener = new CatrobatListener();
       $listener->setIo($container->get('console.io'));
 
